phase 1: narrow types, add StatusFlags, validated setters, fix sentinel
beacon_config: Narrow BeaconConfig fields to precise types (uint8_t,
uint16_t, uint32_t) while preserving 4-byte NVS wire format. Add
StatusFlags struct with pack/unpack, WhatInStatus and StatusMode enums.
Replace config_mut() with 12 validated setters that enforce ranges
(mult_period in {1,2,4,8}, change_interval aligned to %8, etc.).
Validate NVS values on load — reject out-of-range (prevents
mult_period=0 busy-loop from corrupt flash).

beacon_logic: Narrow StatusInput fields (status_flags to uint32_t,
battery_voltage to uint16_t, keys_changes to uint16_t, what_in_status
to uint8_t). Clean up bit extraction to shift-then-mask pattern.

accel_data: Fix sentinel anti-pattern — replace last_sample_={-1,-1,-1}
check with explicit has_previous_ bool flag. -1 is a valid int16_t
accelerometer reading.

Tests will not compile yet (config_mut() removed) — test updates
follow in Phase 3.

+++ src/beacon_logic.hpp
@@ -23,12 +23,12 @@ bool is_key_empty(const uint8_t* key, size_t len);
 /// Input to the status computation (replaces all globals).
 /// See status_flags encoding in beacon_config.hpp for the bitfield layout.
 struct StatusInput {
-    int status_flags;    // Packed 32-bit config (see beacon_config.hpp status_flags encoding)
-    int battery_voltage; // Battery voltage in millivolts (e.g. 3800 = 3.8V)
-    int keys_changes;    // Rolling counter of AirTag key rotations (used by mode 2)
-    int what_in_status;  // Telemetry cycle selector: 0=voltage, 1=accel, 2=temperature
-    uint8_t accel_byte;  // 7-bit movement summary from MovementTracker::compute_accel_byte()
-    int16_t temperature; // Accelerometer die temperature in 0.1C units (e.g. 235 = 23.5C)
+    uint32_t status_flags;    // Packed 32-bit config (see beacon_config.hpp status_flags encoding)
+    uint16_t battery_voltage; // Battery voltage in millivolts (e.g. 3800 = 3.8V)
+    uint16_t keys_changes;    // Rolling counter of AirTag key rotations (used by mode 2)
+    uint8_t what_in_status;   // Telemetry cycle selector: 0=voltage, 1=accel, 2=temperature
+    uint8_t accel_byte;       // 7-bit movement summary from MovementTracker::compute_accel_byte()
+    int16_t temperature;      // Accelerometer die temperature in 0.1C units (e.g. 235 = 23.5C)
 };
 
 /// Output of the status computation.
